changing to float

+++ src/ExternalConnections/Delivery/Delivery.java
@@ -8,7 +8,7 @@ public interface Delivery {
     // 0 is success, 1 is connection invalid, 2 not deliverable.
     // Recieve list of items
     // todo what does delievery recieve.
-    int Delivery (int delivery);
+    int Delivery (float delivery);
 
     // use secure key to connect
     boolean connect (int key);
